simplify parser code
