+++ csv/parsers/types.go
@@ -19,7 +19,7 @@ type ParsingGroup interface {
 	String() string
 	AddTxToGroup(db.TaxableTransaction)
 	GetGroupedTxes() map[uint][]db.TaxableTransaction
-	ParseGroup() error
+	ParseGroup(parsingFunc func(sf *WrapperLpTxGroup) error) error
 	GetRowsForParsingGroup() []CsvRow
 }
